+++ test/Main.hs
@@ -5,8 +5,10 @@ import Control.Monad.IO.Class
 import Data.Either
 import Data.Typeable
 import Data.UUID.Types
+import qualified Data.Set as Set
 import qualified Data.Text as T
 
+import Data.Monoid.Utils
 import Database.PostgreSQL.PQTypes
 import Database.PostgreSQL.PQTypes.Checks
 import Database.PostgreSQL.PQTypes.Model.ColumnType
@@ -16,6 +18,7 @@ import Database.PostgreSQL.PQTypes.Model.Index
 import Database.PostgreSQL.PQTypes.Model.Migration
 import Database.PostgreSQL.PQTypes.Model.PrimaryKey
 import Database.PostgreSQL.PQTypes.Model.Table
+import Database.PostgreSQL.PQTypes.Model.Trigger
 import Database.PostgreSQL.PQTypes.SQL.Builder
 import Log
 import Log.Backend.StandardOutput
@@ -69,6 +72,7 @@ tableBankSchema1 =
                 , colNullable = False }
     ]
   , tblPrimaryKey = pkOnColumn "id"
+  , tblTriggers = []
   }
 
 tableBankSchema2 :: Table
@@ -368,7 +372,8 @@ schema2Migrations :: (MonadDB m) => [Migration m]
 schema2Migrations = schema1Migrations
                  ++ [ dropTableMigration   tableWitnessedRobberySchema1
                     , dropTableMigration   tableWitnessSchema1
-                    , createTableMigration tableUnderArrestSchema2 ]
+                    , createTableMigration tableUnderArrestSchema2
+                    ]
 
 schema3Tables :: [Table]
 schema3Tables = [ tableBankSchema3
@@ -826,6 +831,325 @@ migrationTest1Body step = do
   migrateDBToSchema5  step
   testDBSchema5       step
 
+bankTrigger1 :: Trigger
+bankTrigger1 =
+  Trigger { triggerTable = "bank"
+          , triggerName = "trigger_1"
+          , triggerEvents = Set.fromList [TriggerInsert]
+          , triggerDeferrable = False
+          , triggerInitiallyDeferred = False
+          , triggerWhen = Nothing
+          , triggerFunction = TriggerFunction "function_1" $
+                "begin"
+            <+> "  perform true;"
+            <+> "  return null;"
+            <+> "end;"
+          }
+
+bankTrigger2 :: Trigger
+bankTrigger2 =
+  bankTrigger1
+  { triggerFunction = TriggerFunction "function_2" $
+          "begin"
+      <+> "  return null;"
+      <+> "end;"
+  }
+
+bankTrigger3 :: Trigger
+bankTrigger3 =
+  Trigger { triggerTable = "bank"
+          , triggerName = "trigger_3"
+          , triggerEvents = Set.fromList [TriggerInsert, TriggerUpdateOf [unsafeSQL "location"]]
+          , triggerDeferrable = True
+          , triggerInitiallyDeferred = True
+          , triggerWhen = Nothing
+          , triggerFunction = TriggerFunction "function_3" $
+                "begin"
+            <+> "  perform true;"
+            <+> "  return null;"
+            <+> "end;"
+          }
+
+bankTrigger2Proper :: Trigger
+bankTrigger2Proper =
+  bankTrigger2 { triggerName = "trigger_2" }
+
+testTriggers :: HasCallStack => (String -> TestM ()) -> TestM ()
+testTriggers step = do
+  step "Running trigger tests..."
+
+  step "create the initial database"
+  migrate [tableBankSchema1] [createTableMigration tableBankSchema1]
+
+  do
+    let msg = "checkDatabase fails if there are triggers in the database but not in the schema"
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = []
+                                }
+             ]
+        ms = [ createTriggerMigration 1 bankTrigger1 ]
+    step msg
+    assertException msg $ migrate ts ms
+
+  do
+    let msg = "checkDatabase fails if there are triggers in the schema but not in the database"
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [bankTrigger1]
+                                }
+             ]
+        ms = []
+    triggerStep msg $ do
+      assertException msg $ migrate ts ms
+
+  do
+    let msg = "test succeeds when creating a single trigger"
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [bankTrigger1]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 bankTrigger1 ]
+    triggerStep msg $ do
+      assertNoException msg $ migrate ts ms
+      verify [bankTrigger1] True
+
+  do
+    let msg = "checkDatabase fails if triggers differ in function name"
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [bankTrigger1]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 bankTrigger2 ]
+    triggerStep msg $ do
+      assertException msg $ migrate ts ms
+
+  do
+    -- Attempt to create the same triggers twice. Should fail with a DBException saying
+    -- that function already exists.
+    let msg = "database exception is raised if trigger is created twice"
+        ts = [ tableBankSchema1 { tblVersion = 3
+                                , tblTriggers = [bankTrigger1]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 bankTrigger1
+             , createTriggerMigration 2 bankTrigger1
+             ]
+    triggerStep msg $ do
+      assertDBException msg $ migrate ts ms
+
+  do
+    let msg = "database exception is raised if triggers only differ in function name"
+        ts = [ tableBankSchema1 { tblVersion = 3
+                                , tblTriggers = [bankTrigger1, bankTrigger2]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 bankTrigger1
+             , createTriggerMigration 2 bankTrigger2
+             ]
+    triggerStep msg $ do
+      assertDBException msg $ migrate ts ms
+
+  do
+    let msg = "successfully migrate two triggers"
+        ts = [ tableBankSchema1 { tblVersion = 3
+                                , tblTriggers = [bankTrigger1, bankTrigger2Proper]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 bankTrigger1
+             , createTriggerMigration 2 bankTrigger2Proper
+             ]
+    triggerStep msg $ do
+      assertNoException msg $ migrate ts ms
+      verify [bankTrigger1, bankTrigger2Proper] True
+
+  do
+    let msg = "database exception is raised if trigger's WHEN is syntactically incorrect"
+        trg = bankTrigger1 { triggerWhen = Just "WILL FAIL" }
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 trg ]
+    triggerStep msg $ do
+      assertDBException msg $ migrate ts ms
+
+  do
+    let msg = "database exception is raised if trigger's WHEN uses undefined column"
+        trg = bankTrigger1 { triggerWhen = Just "NEW.foobar = 1" }
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 trg ]
+    triggerStep msg $ do
+      assertDBException msg $ migrate ts ms
+
+  do
+    -- This trigger is valid. However, the WHEN clause specified in triggerWhen is not
+    -- what gets returned from the database. The decompiled and normalized WHEN clause
+    -- from the database looks like this:
+    --   new.name <> 'foobar'::text
+    -- We simply assert an exception, which presumably comes from the migration framework,
+    -- while it should actually be a deeper check for just the differing WHEN
+    -- clauses. On the other hand, it's probably good enough as it is.
+    -- See the comment for 'getDBTriggers' in src/Database/PostgreSQL/PQTypes/Model/Trigger.hs.
+    let msg = "checkDatabase fails if WHEN clauses from database and code differ"
+        trg = bankTrigger1 { triggerWhen = Just "NEW.name != 'foobar'" }
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 trg ]
+    triggerStep msg $ do
+      assertException msg $ migrate ts ms
+
+  do
+    let msg = "successfully migrate trigger with valid WHEN"
+        trg = bankTrigger1 { triggerWhen = Just "new.name <> 'foobar'::text" }
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 trg ]
+    triggerStep msg $ do
+      assertNoException msg $ migrate ts ms
+      verify [trg] True
+
+  do
+    let msg = "successfully migrate trigger that is deferrable"
+        trg = bankTrigger1 { triggerDeferrable = True }
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 trg ]
+    triggerStep msg $ do
+      assertNoException msg $ migrate ts ms
+      verify [trg] True
+
+  do
+    let msg = "successfully migrate trigger that is deferrable and initially deferred"
+        trg = bankTrigger1 { triggerDeferrable = True
+                           , triggerInitiallyDeferred = True
+                           }
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 trg ]
+    triggerStep msg $ do
+      assertNoException msg $ migrate ts ms
+      verify [trg] True
+
+  do
+    let msg = "database exception is raised if trigger is initially deferred but not deferrable"
+        trg = bankTrigger1 { triggerDeferrable = False
+                           , triggerInitiallyDeferred = True
+                           }
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 trg ]
+    triggerStep msg $ do
+      assertDBException msg $ migrate ts ms
+
+  do
+    let msg = "database exception is raised if dropping trigger that does not exist"
+        trg = bankTrigger1
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ dropTriggerMigration 1 trg ]
+    triggerStep msg $ do
+      assertDBException msg $ migrate ts ms
+
+  do
+    let msg = "database exception is raised if dropping trigger function of which does not exist"
+        trg = bankTrigger2
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ dropTriggerMigration 1 trg ]
+    triggerStep msg $ do
+      assertDBException msg $ migrate ts ms
+
+  do
+    let msg = "successfully drop trigger"
+        trg = bankTrigger1
+        ts = [ tableBankSchema1 { tblVersion = 3
+                                , tblTriggers = []
+                                }
+             ]
+        ms = [ createTriggerMigration 1 trg, dropTriggerMigration 2 trg ]
+    triggerStep msg $ do
+      assertNoException msg $ migrate ts ms
+      verify [trg] False
+
+  do
+    let msg = "database exception is raised if dropping trigger twice"
+        trg = bankTrigger2
+        ts = [ tableBankSchema1 { tblVersion = 3
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ dropTriggerMigration 1 trg, dropTriggerMigration 2 trg ]
+    triggerStep msg $ do
+      assertDBException msg $ migrate ts ms
+
+  do
+    let msg = "successfully create trigger with multiple events"
+        trg = bankTrigger3
+        ts = [ tableBankSchema1 { tblVersion = 2
+                                , tblTriggers = [trg]
+                                }
+             ]
+        ms = [ createTriggerMigration 1 trg ]
+    triggerStep msg $ do
+      assertNoException msg $ migrate ts ms
+      verify [trg] True
+
+  where
+    triggerStep msg rest = do
+      recreateTriggerDB
+      step msg
+      rest
+
+    migrate tables migrations = do
+      migrateDatabase defaultExtrasOptions ["pgcrypto"] [] [] tables migrations
+      checkDatabase defaultExtrasOptions [] [] tables
+
+    -- Verify that the given triggers are (not) present in the database.
+    verify :: (MonadIO m, MonadDB m, HasCallStack) => [Trigger] -> Bool -> m ()
+    verify triggers present = do
+      dbTriggers <- getDBTriggers "bank"
+      let ok = and $ map (`elem` dbTriggers) triggers
+          err = "Triggers " <> (if present then "" else "not ") <> "present in the database."
+          trans = if present then id else not
+      liftIO . assertBool err $ trans ok
+
+    triggerMigration :: MonadDB m => (Trigger -> m ()) -> Int -> Trigger -> Migration m
+    triggerMigration fn from trg = Migration
+      { mgrTableName = tblName tableBankSchema1
+      , mgrFrom = fromIntegral from
+      , mgrAction = StandardMigration $ fn trg
+      }
+
+    createTriggerMigration :: MonadDB m => Int -> Trigger -> Migration m
+    createTriggerMigration = triggerMigration createTrigger
+
+    dropTriggerMigration :: MonadDB m => Int -> Trigger -> Migration m
+    dropTriggerMigration = triggerMigration dropTrigger
+
+    recreateTriggerDB = do
+      runSQL_ "DROP TRIGGER IF EXISTS trg__bank__trigger_1 ON bank;"
+      runSQL_ "DROP TRIGGER IF EXISTS trg__bank__trigger_2 ON bank;"
+      runSQL_ "DROP FUNCTION IF EXISTS function_1;"
+      runSQL_ "DROP FUNCTION IF EXISTS function_2;"
+      runSQL_ "DROP TABLE IF EXISTS bank;"
+      runSQL_ "DELETE FROM table_versions WHERE name = 'bank'";
+      migrate [tableBankSchema1] [createTableMigration tableBankSchema1]
 
 migrationTest1 :: ConnectionSourceM (LogT IO) -> TestTree
 migrationTest1 connSource =
@@ -833,8 +1157,6 @@ migrationTest1 connSource =
   freshTestDB         step
 
   migrationTest1Body  step
-
-  -- freshTestDB         step
 
 -- | Test for behaviour of 'checkDatabase' and 'checkDatabaseAllowUnknownObjects'
 migrationTest2 :: ConnectionSourceM (LogT IO) -> TestTree
@@ -957,6 +1279,13 @@ migrationTest4 connSource =
 
   freshTestDB         step
 
+-- | Test triggers.
+triggerTests :: ConnectionSourceM (LogT IO) -> TestTree
+triggerTests connSource =
+  testCaseSteps' "Trigger tests" connSource $ \step -> do
+    freshTestDB  step
+    testTriggers step
+
 eitherExc :: MonadCatch m => (SomeException -> m ()) -> (a -> m ()) -> m a -> m ()
 eitherExc left right c = try c >>= either left right
 
@@ -966,9 +1295,14 @@ assertNoException t c = eitherExc
   (const $ return ()) c
 
 assertException :: String -> TestM () -> TestM ()
-assertException   t c = eitherExc
+assertException t c = eitherExc
   (const $ return ())
   (const $ liftIO $ assertFailure ("No exception thrown for: " ++ t)) c
+
+assertDBException :: String -> TestM () -> TestM ()
+assertDBException t c =
+  try c >>= either (\DBException{} -> pure ())
+                   (const . liftIO . assertFailure $ "No DBException thrown for: " ++ t)
 
 -- | A variant of testCaseSteps that works in TestM monad.
 testCaseSteps' :: TestName -> ConnectionSourceM (LogT IO)
@@ -994,6 +1328,7 @@ main = do
                          , migrationTest2 connSource
                          , migrationTest3 connSource
                          , migrationTest4 connSource
+                         , triggerTests connSource
                          ]
   where
     ings =